fix ensureDefaultConfigFile behaviour

+++ pkg/steampipeconfig/load_config.go
@@ -1,11 +1,13 @@
 package steampipeconfig
 
 import (
+	"bytes"
 	"fmt"
 	"log"
 	"os"
 	"path/filepath"
 	"strings"
+	"time"
 
 	"github.com/gertd/go-pluralize"
 	"github.com/hashicorp/hcl/v2"
@@ -50,32 +52,60 @@ func LoadConnectionConfig() (*SteampipeConfig, *error_helpers.ErrorAndWarnings) 
 }
 
 func ensureDefaultConfigFile(configFolder string) error {
-	var equal bool
 	defaultConfigFile := filepath.Join(configFolder, defaultConfigFileName)
 	defaultConfigSampleFile := filepath.Join(configFolder, defaultConfigSampleFileName)
 
-	// always write the default.spc.sample file
-	err := os.WriteFile(defaultConfigSampleFile, []byte(constants.DefaultConnectionConfigContent), 0755)
+	// check if sample and default files exist
+	sampleExists := filehelpers.FileExists(defaultConfigSampleFile)
+	defaultExists := filehelpers.FileExists(defaultConfigFile)
+
+	// case: neither sample nor default.spc exists - write both
+	if !sampleExists {
+		err := os.WriteFile(defaultConfigSampleFile, []byte(constants.DefaultConnectionConfigContent), 0755)
+		if err != nil {
+			return err
+		}
+	}
+	if !defaultExists {
+		err := os.WriteFile(defaultConfigFile, []byte(constants.DefaultConnectionConfigContent), 0755)
+		if err != nil {
+			return err
+		}
+	}
+
+	// get the file infos
+	defaultStat, err := os.Stat(defaultConfigFile)
 	if err != nil {
 		return err
 	}
-	// if default.spc exists, and if the content of the default.spc and the default.spc.sample
-	// file are the same, remove the default.spc file.
-	if _, err := os.Stat(defaultConfigFile); err == nil {
-		if equal, err = utils.AreFilesEqual(defaultConfigFile, defaultConfigSampleFile); err != nil {
+	sampleStat, err := os.Stat(defaultConfigSampleFile)
+	if err != nil {
+		return err
+	}
+	sampleContent, err := os.ReadFile(defaultConfigSampleFile)
+	if err != nil {
+		return err
+	}
+
+	// get the file mod times
+	defaultModTime := defaultStat.ModTime()
+	sampleModTime := sampleStat.ModTime()
+
+	// check if the files are modified
+	defaultModified := defaultModTime.After(sampleModTime) && defaultModTime.Sub(sampleModTime) > 100*time.Millisecond
+	sampleModified := !bytes.Equal([]byte(constants.DefaultConnectionConfigContent), sampleContent)
+
+	// case: if sample is modified - always write new sample file content
+	if sampleModified {
+		err := os.WriteFile(defaultConfigSampleFile, []byte(constants.DefaultConnectionConfigContent), 0755)
+		if err != nil {
 			return err
 		}
-		if equal {
-			// remove default.spc file
-			err := os.Remove(defaultConfigFile)
-			if err != nil {
-				return err
-			}
-		}
 	}
-	// write default.spc if it does not exist
-	if _, err := os.Stat(defaultConfigFile); os.IsNotExist(err) {
-		err = os.WriteFile(defaultConfigFile, []byte(constants.DefaultConnectionConfigContent), 0755)
+
+	// case: if sample is modified but default is not modified - write the new default file content
+	if sampleModified && !defaultModified {
+		err := os.WriteFile(defaultConfigFile, []byte(constants.DefaultConnectionConfigContent), 0755)
 		if err != nil {
 			return err
 		}
